Adjust the UAAGO library
We'd like to provide the nozzle app to our users. In order to make it
simplish, we'd like to allow the app to use username and password to
authenticate as oppose to the client_credentials. This initially threw
some problems, returning 401 constantly - which we've managed to track
down and fix by setting different Basic Auth credentials.